services/student-coach-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-dotenv==1.0.0
requests==2.32.3
python-multipart==0.0.20
psycopg2-binary==2.9.10

